refactor: movendo página de Chat para subfeature
