+++ frontend/src/features/services/execution/components/Chat/LimitadorLargura.tsx
