+++ frontend/src/features/services/execution/hooks/useChatConnection.tsx
