+++ frontend/src/features/services/execution/routes/executionPaths.ts
@@ -1,0 +1,4 @@
+export const executionPaths = {
+  chatServico: "/chat-servico",
+  chatServicoById: (id: string | number) => `/chat-servico/${id}`,
+};
