+++ frontend/src/features/services/routes/servico.routes.tsx
@@ -4,18 +4,9 @@ import PesquisaServico from "@/features/services/pages/PesquisaServicos";
 import { VisualizarServico } from "@/features/services/pages/VisualizarServico";
 import { MeuServico } from "@/features/services/pages/MeuServico";
 import { servicoPaths } from "@/features/services/routes/servicoPaths";
-import { ChatServico } from "../pages/ChatServico";
 
 export const servicoRoutes = [
   { path: servicoPaths.pesquisaServico, element: <PesquisaServico /> },
-  {
-    path: `${servicoPaths.chatServico}/:id`,
-    element: (
-      <RequireRole roles={["Cliente", "Freelancer"]}>
-        <ChatServico />
-      </RequireRole>
-    ),
-  },
 
   {
     path: `${servicoPaths.visualizarServico}/:id`,
