+++ frontend/src/features/services/routes/servicoPaths.ts
@@ -3,9 +3,7 @@ export const servicoPaths = {
   visualizarServico: "/visualizar-servico",
   cadastrarServico: "/cadastrar-servico",
   meuServico: "/meu-servico",
-  chatServico: "/chat-servico",
 
   visualizarServicoById: (id: string | number) => `/visualizar-servico/${id}`,
   visualizarMeuServicoById: (id: string | number) => `/meu-servico/${id}`,
-  chatServicoById: (id: string | number) => `/chat-servico/${id}`,
 };
